feat: Add movie by genre page and refactor movie view component
This commit adds a new page for movies by genre and refactors the movie view component to be located in the medias module instead of the social-network module. The necessary changes were made to routing modules, navigation links, and module declarations. A new card module was also imported into the medias module.
