+++ src/app/pages/medias/medias-routing.module.ts
@@ -3,10 +3,11 @@ import { RouterModule, Routes } from '@angular/router';
 import { NavigationLinks, navigationRoot } from '../../app-routing.module';
 import { MediasMediaListComponent } from './medias-media-list/medias-media-list.component';
 import { TrendingComponent } from './trending/trending.component';
-import { MovieViewComponent } from '../social-network/movie-view/movie-view.component';
+import { MovieViewComponent } from './movie-view/movie-view.component';
+import { MovieByGenreComponent } from './movie-by-genre/movie-by-genre.component';
 
 export const mediasLinks: NavigationLinks<
-  'list' | 'search' | 'trending' | 'movie_view'
+  'list' | 'search' | 'trending' | 'movie_view' | 'movies_by_genre'
 > = {
   list: {
     path: 'list',
@@ -24,10 +25,17 @@ export const mediasLinks: NavigationLinks<
     path: 'movie-view',
     name: 'Film',
   },
+  movies_by_genre: {
+    path: 'movies-by-genre',
+    name: 'Films par genre',
+  },
 };
 
 export const movieViewPath =
   '/' + navigationRoot.medias.path + '/' + mediasLinks.movie_view.path;
+
+export const moviesByGenrePath =
+  '/' + navigationRoot.medias.path + '/' + mediasLinks.movies_by_genre.path;
 
 const routes: Routes = [
   {
@@ -49,6 +57,11 @@ const routes: Routes = [
     path: mediasLinks.movie_view.path + '/:id',
     component: MovieViewComponent,
     data: { title: mediasLinks.movie_view.name },
+  },
+  {
+    path: mediasLinks.movies_by_genre.path + '/:id',
+    component: MovieByGenreComponent,
+    data: { title: mediasLinks.movies_by_genre.name },
   },
 ];
