+++ src/app/pages/medias/movie-view/movie-view.component.ts
